separate bitrise related endpoints
